move celery stuff to own directory, pin Pipfile
