+++ Pipfile
@@ -4,42 +4,45 @@ url = "https://pypi.org/simple"
 verify_ssl = true
 
 [dev-packages]
-django-extensions = "*"
-werkzeug = "*"
-django-test-plus = "*"
-coverage = "*"
-django-coverage-plugin = "*"
-ipython = "*"
-ipdb = "*"
-pytest-django = "*"
-pytest-sugar = "*"
-pytest-cov = "*"
-codecov = "*"
-freezegun = "*"
-pylint = "*"
-pylint-django = "*"
-flake8 = "*"
-isort = "*"
-black = "*"
+django-extensions = "==2.1.4"
+Werkzeug = "==0.14.1"
+django-test-plus = "==1.1.1"
+coverage = "==5.0a4"
+django_coverage_plugin = "==1.6.0"
+ipython = "==7.2.0"
+ipdb = "==0.11"
+pytest-django = "==3.4.4"
+pytest-sugar = "==0.9.2"
+pytest-cov = "==2.6.1"
+codecov = "==2.0.15"
+freezegun = "==0.3.11"
+pylint = "==2.2.2"
+pylint-django = "==2.0.4"
+flake8 = "==3.6.0"
+isort = "==4.3.4"
+black = "==18.9b0"
 
 [packages]
-django = "*"
-gunicorn = "*"
-gevent = "*"
-raven = "*"
-whitenoise = "*"
-django-guardian = "*"
-django-environ = "*"
-djoser = "*"
-python-dateutil = "*"
-django-anymail = "*"
-django-taggit = "*"
-psycopg2 = "*"
-psycopg2-binary = "*"
-djangorestframework-simplejwt = "*"
-djangorestframework = "*"
-celery = "*"
-drf-yasg = "*"
+Django = "==2.1.5"
+gunicorn = "==19.9.0"
+gevent = "==1.4.0"
+raven = "==6.10.0"
+whitenoise = "==4.1.2"
+django-guardian = "==1.4.9"
+django-environ = "==0.4.5"
+djoser = "==1.5.1"
+python-dateutil = "==2.7.5"
+django-anymail = "==5.0"
+django-taggit = "==0.23.0"
+psycopg2 = "==2.7.7"
+psycopg2-binary = "==2.7.7"
+djangorestframework = "==3.9.1"
+djangorestframework-simplejwt = "==3.3"
+celery = "==4.2.1"
+drf-yasg = "==1.12.1"
 
 [requires]
 python_version = "3.6"
+
+[pipenv]
+allow_prereleases = true
