+++ docker/local/Dockerfile.celery
@@ -7,4 +7,4 @@ COPY Pipfile Pipfile.lock /
 RUN pip install pipenv
 RUN LIBRARY_PATH=/lib:/usr/lib /bin/sh -c "pipenv install --dev --system"
 RUN useradd -ms /bin/bash celery-user
-ENTRYPOINT celery worker -A api.celery --concurrency=3 --loglevel=WARNING --uid=celery-user
+ENTRYPOINT celery worker -A project_celery.celery --concurrency=3 --loglevel=WARNING --uid=celery-user
